Refactored component names to CapitalCase
