+++ canvas_fe/src/App.js
@@ -1,6 +1,6 @@
 import React, {Component, Fragment} from "react";
-import Header from "./components/header";
-import Home from "./components/home";
+import Header from "./components/Header";
+import Home from "./components/Home";
 
 
 class App extends Component {
